+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 <a name="unreleased"></a>
 ## [Unreleased]
 
+
+<a name="v2.4.0"></a>
+## [v2.4.0] - 2021-12-19
 ### Feat
 - **golang:** moved from context organize imports to goimports in null ls
 - **null_ls:** added sqlformat and codespell diagnostics
@@ -355,7 +358,8 @@
 - **indentation:** moved indentation to it's own folder
 
 
-[Unreleased]: https://github.com/tigorlazuardi/nvim/compare/v2.3.1...HEAD
+[Unreleased]: https://github.com/tigorlazuardi/nvim/compare/v2.4.0...HEAD
+[v2.4.0]: https://github.com/tigorlazuardi/nvim/compare/v2.3.1...v2.4.0
 [v2.3.1]: https://github.com/tigorlazuardi/nvim/compare/v2.3.0...v2.3.1
 [v2.3.0]: https://github.com/tigorlazuardi/nvim/compare/v2.2.0...v2.3.0
 [v2.2.0]: https://github.com/tigorlazuardi/nvim/compare/v2.0.3...v2.2.0